requirements.txt
python-telegram-bot==13.15
google-auth==2.15.0
google-auth-oauthlib==0.7.2
google-auth-httplib2==0.1.0
google-api-python-client==2.70.0
openai==0.27.0
requests==2.28.1